+++ client/src/components/DisplayTypeContainer/DisplayTypeContainer.js
