--- client/src/components/libraries/Following.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Following() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
-
-export default Following;
