--- client/src/components/libraries/Subscriptions.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Subscriptions() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
-
-export default Subscriptions
